# TaxDeedFlow/public/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tax Deed Flow API",
    "version": "1.0.0",
    "description": "Public REST API for accessing tax deed property data, county information, and risk analysis. Requires API key authentication.",
    "contact": {
      "name": "Tax Deed Flow Support",
      "url": "https://taxdeedflow.com/developer"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://taxdeedflow.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/properties/{id}": {
      "get": {
        "tags": ["Properties"],
        "summary": "Get property details",
        "description": "Retrieve detailed information about a specific property",
        "operationId": "getProperty",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Property ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPropertyResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitExceeded"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for accessing public v1 endpoints. Generate keys at /developer/keys"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session-based authentication for developer management endpoints"
      }
    },
    "schemas": {
      "Property": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Property unique identifier"
          },
          "county_id": {
            "type": "string",
            "format": "uuid",
            "description": "County identifier"
          },
          "parcel_id": {
            "type": "string",
            "nullable": true,
            "description": "Property parcel number"
          },
          "address": {
            "type": "string",
            "nullable": true,
            "description": "Property address"
          },
          "total_due": {
            "type": "number",
            "nullable": true,
            "description": "Total amount owed"
          },
          "sale_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Scheduled auction date"
          },
          "investability_score": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "maximum": 100,
            "description": "AI-calculated investability score (0-100)"
          },
          "has_regrid_data": {
            "type": "boolean",
            "description": "Whether Regrid property data is available"
          },
          "has_screenshot": {
            "type": "boolean",
            "description": "Whether aerial/street view images are available"
          },
          "visual_validation_status": {
            "type": "string",
            "enum": ["approved", "rejected", "caution", "pending"],
            "nullable": true,
            "description": "Visual validation status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Record creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Record last updated timestamp"
          }
        }
      },
      "PropertyWithDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Property"
          },
          {
            "type": "object",
            "properties": {
              "county": {
                "type": "object",
                "nullable": true,
                "properties": {
                  "county_name": {
                    "type": "string"
                  },
                  "state_code": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ]
      },
...
      "GetPropertyResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PropertyWithDetails"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized - missing or invalid API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Unauthorized",
              "message": "Missing or invalid API key"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Forbidden",
              "message": "Your API key does not have read permission"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found - resource does not exist",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Not Found",
              "message": "Property not found"
            }
          }
        }
      },
      "RateLimitExceeded": {
        "description": "Rate limit exceeded",
        "headers": {
          "X-RateLimit-Limit": {
            "description": "The rate limit ceiling for this endpoint",
            "schema": {
              "type": "integer"
            }
          },
          "X-RateLimit-Remaining": {
            "description": "The number of requests remaining in the current window",
            "schema": {
              "type": "integer"
            }
          },
          "X-RateLimit-Reset": {
            "description": "Unix timestamp when the rate limit resets",
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Rate Limit Exceeded",
              "message": "Rate limit exceeded. Limit: 1000/hour. Resets at 2024-01-23T15:00:00Z"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Server Error",
              "message": "An unexpected error occurred"
            }
          }
        }
      }
    }
  }
}